@@ -14,43 +14,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.6",
-    "@angular/cdk": "^9.2.3",
-    "@angular/common": "^9.1.6",
-    "@angular/compiler": "^9.1.6",
-    "@angular/core": "^9.1.6",
-    "@angular/forms": "^9.1.6",
-    "@angular/http": "^7.2.16",
-    "@angular/material": "^9.2.3",
-    "@angular/platform-browser": "^9.1.6",
-    "@angular/platform-browser-dynamic": "^9.1.6",
-    "@angular/router": "^9.1.6",
+    "@angular/animations": "^10.1.4",
+    "@angular/cdk": "^10.2.4",
+    "@angular/common": "^10.1.4",
+    "@angular/compiler": "^10.1.4",
+    "@angular/core": "^10.1.4",
+    "@angular/forms": "^10.1.4",
+    "@angular/material": "^10.2.4",
+    "@angular/platform-browser": "^10.1.4",
+    "@angular/platform-browser-dynamic": "^10.1.4",
+    "@angular/router": "^10.1.4",
     "core-js": "^3.6.5",
-    "hammerjs": "^2.0.8",
-    "rxjs": "^6.5.3",
-    "zone.js": "^0.10.2"
+    "rxjs": "^6.6.3",
+    "zone.js": "^0.11.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.901.5",
-    "@angular/cli": "^9.1.5",
-    "@angular/compiler-cli": "^9.1.6",
-    "@types/jasmine": "~3.5.10",
-    "@types/node": "~13.13.5",
-    "codelyzer": "~5.2.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.2",
-    "karma": "^5.0.5",
+    "@angular-devkit/build-angular": "^0.1001.4",
+    "@angular/cli": "^10.1.4",
+    "@angular/compiler-cli": "^10.1.4",
+    "@types/jasmine": "~3.5.14",
+    "@types/node": "~14.11.5",
+    "codelyzer": "~6.0.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~6.0.0",
+    "karma": "~5.2.3",
     "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^3.0.2",
-    "karma-jasmine": "~3.1.1",
-    "karma-jasmine-html-reporter": "^1.5.3",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "phantomjs-prebuilt": "^2.1.16",
-    "prettier": "^2.0.5",
-    "protractor": "^6.0.0",
-    "ts-node": "~8.10.1",
-    "tslint": "~6.1.2",
-    "typescript": "~3.8.3"
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "prettier": "^2.1.2",
+    "protractor": "~7.0.0",
+    "ts-node": "~9.0.0",
+    "tslint": "~6.1.3",
+    "typescript": "~3.9.2"
   }
 }
